requirement.txt
Django==2.0.5
django-autoslug==1.9.3
django-bootstrap-form==3.4
django-bootstrap-select==0.1.3
django-bootstrap3==10.0.1
django-braces==1.13.0
django-easy-select2==1.5.4
django-filter==2.0.0
django-safedelete==0.5.1
django-softdelete==0.8.0
django-softdelete-it==0.1
django-tags-input==4.1.0
django-widget-tweaks==1.4.2
djangodeletes==0.43
djangorestframework==3.8.2
Faker==0.9.0
gunicorn==19.9.0
links-collection==0.1.9
Markdown==2.6.11
Pillow==5.1.0
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
slug==1.0
sorl-thumbnail==12.4.1
South==1.0.2
text-unidecode==1.2
uuid==1.30
whitenoise==4.1
